Conflicts:
	NEWS

+++ libsigrok4DSL/libsigrok-internal.h
@@ -83,7 +83,7 @@ SR_PRIV int sr_err(const char *format, ...);
 
 /*--- device.c --------------------------------------------------------------*/
 
-SR_PRIV struct sr_channel *sr_channel_new(int index, int type,
+SR_PRIV struct sr_channel *sr_channel_new(uint16_t index, int type,
                                           gboolean enabled, const char *name);
 SR_PRIV void sr_dev_probes_free(struct sr_dev_inst *sdi);
